# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/community/{route}/member-requests": {
            "get": {
                "tags": [
                    "Community's Member Request"
                ],
                "summary": "Get community's member request by route",
                "operationId": "b95cb3f91a15d33299234611af2d5229",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route of the community",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1mmee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Community member requests successfully"
                    },
                    "500": {
                        "description": "Cannot get Community member requests!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/community/{route}/members": {
            "get": {
                "tags": [
                    "Community's Member"
                ],
                "summary": "Get community's member by route",
                "operationId": "7903776ddbe33134c9a8db31baa5b316",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route of the community",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1mmee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Community members successfully"
                    },
                    "500": {
                        "description": "Cannot get Community members!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}